Should be fine.

+++ ddl/postgres/deapp/de_subject_protein_data.sql
@@ -4,7 +4,7 @@
 CREATE TABLE de_subject_protein_data (
     trial_name character varying(15),
     component character varying(15),
-    intensity bigint,
+    intensity numeric,
     patient_id bigint,
     subject_id character varying(10),
     gene_symbol character varying(100),
@@ -12,9 +12,9 @@ CREATE TABLE de_subject_protein_data (
     assay_id bigint,
     timepoint character varying(20),
     n_value bigint,
-    mean_intensity bigint,
-    stddev_intensity bigint,
-    median_intensity bigint,
-    zscore bigint
+    mean_intensity numeric,
+    stddev_intensity numeric,
+    median_intensity numeric,
+    zscore double precision
 );
